docs : Create folder for static data in Footer.js and move data
