--- src/component/Footer/static.js
@@ -1,20 +1,0 @@
-import Home from '../../assets/images/footer/footer_reponsive/Home_gray.svg'
-import ClickedHome from '../../assets/images/footer/footer_reponsive/home.svg'
-import Review from '../../assets/images/footer/footer_reponsive/Review_gray.svg'
-import ClickedReview from '../../assets/images/footer/footer_reponsive/Review.svg'
-import LectureMaterial from '../../assets/images/footer/footer_reponsive/LectureMaterial_gray.svg'
-import ClickedLectureMaterial from '../../assets/images/footer/footer_reponsive/LectureMaterial.svg'
-import Timetable from '../../assets/images/footer/footer_reponsive/Timetable_gray.svg'
-import ClickedTimetable from '../../assets/images/footer/footer_reponsive/Timetable.svg'
-import MyPage from '../../assets/images/footer/footer_reponsive/MyPage_gray.svg'
-import ClickedMyPage from '../../assets/images/footer/footer_reponsive/MyPage.svg'
-
-export const UnclickedImageList = [Home,Review,LectureMaterial,Timetable,MyPage]
-export const ClickedImageList = [ClickedHome,ClickedReview,ClickedLectureMaterial,ClickedTimetable,ClickedMyPage]
-export const linkList = [
-    'https://bcsdlab.com/',
-    'https://www.koreatech.ac.kr/kor/Main.do',
-    'https://portal.koreatech.ac.kr/login.jsp',
-    '/',
-    'https://www.facebook.com/bcsdlab/'
-    ]
