+++ src/static/footer/menuList.js
@@ -1,0 +1,8 @@
+export const FOOTER_MENU_LIST = ['BCSD Lab 바로가기','코리아텍 바로가기','아우누리 바로가기','개인정보 처리방침']
+export const FOOTER_LINK_LIST = [
+    'https://bcsdlab.com/',
+    'https://www.koreatech.ac.kr/kor/Main.do',
+    'https://portal.koreatech.ac.kr/login.jsp',
+    '/',
+    'https://www.facebook.com/bcsdlab/'
+]
