--- src/static/indexPage/menuList.js
@@ -1,4 +1,0 @@
-export const MENU_LIST = ['홈','강의평','강의자료','시간표']
-export const FOOTER_MENU_LIST = ['BCSD Lab 바로가기','코리아텍 바로가기','아우누리 바로가기','개인정보 처리방침']
-export const MOBILE_MENU_LIST = ['홈','강의평','강의자료','시간표','마이페이지']
-export const MENU_LINK_LIST = ['/','/dummy','/dummy','/dummy']
